Added a Reviews Page
A new Review Page has been added as well as the addition of a Navigation Bar.
The Review page contains Reviews about Games.
The Navigation bar contains a "sticky" feature when it reaches the top of the screen.
The Top Text now contains a background colour and a double border.