New Review! + Updated Review Page & Footer
Added a new Review for Necropolis.
Now when you hover over the Name of a Game it changes colour and becomes underlined.
The Review page has been changed so the Quote now has a dashed box around it.
The border around the Review Text has been removed.
A footer has been added to every page with a discord link if a user wishes to report any bugs.
The Navigation bar recieved a black border underneath.
The Footer recieved a black bar on top.
The first page (index) now displays the icon on the tab.
Fixed the link on the home page to the original Website.