Added Text Font and Colour Palette Change
The Website now has a general Font for all of it's Text to remove some of the blandness when looking at it.
The Colout Palette for the Website has changed from the White and Blue, which in my opinion made it look like a big advert for some Greek/Gyros Restaurant to a bit of a Darker (Rusty?) Overall Theme.
Also added <header> Tags and <nav> Tags for the corresponding elements to help with the A11Y (Accesibility) Project.
A New Review for Rainbow Six Siege has been written!!
Review Scores now have a background-color to make them more impactful.