Added A - Z
A complete list from A - Z has been added on the Music Page to allow quick navigation within the page.
Added a handfull of New Songs to complete the alphabet. (Still missing a song for Q, U and X
Added a smooth scrolling feature when clicking on internal links.
Added a <main> tag to each file.