Massive Update on About Page
The Entire About Page has been filled with content now and gives a little backstory into the Website and the lives of the four people running Touch Digital.
This massive change is one of the biggest updates I've made so far, so be sure to check out the About Page next time!
The Index page has also been updated with new information based on the About Section.