Added a New Song & Review
Added Driving Home for Christmas by Chris Rea.
Added a review for Minecraft on the Review page.
Removed all of the "Back to Top" links, due to them being replaced by the button on the bottom right.
Removed the colour change when hovering over a title on the review page.
Fixed the Absolver image alt attribute.