Complete Revision of the Home page
The Home page has recieved a complete changeover and now features the newest reviews, pictures and videos. (Fine tuning this one took all day.)
Moved every single Image to a new Images folder to tidy the folder up a bit.
The notice just above every footer has been updated to fit with the group of friends story.
Fixed the navbar so it stacks vertically when the screen gets too small.
Adjusted some of the text on the very first page.