Complete Revision of the Home page
The Home page has recieved a complete changeover and now features reviews and pictures. (Fine tuning this one took two days.)
Moved every single Image to a new Images folder to tidy the folder up a bit.
The notice just above every footer has been updated to fit with the group of friends story.
Fixed the navbar so it stacks vertically when the screen gets too small.
Added internal links to allow the elements on the first page link to specific areas on other pages.
Adjusted some of the text on the very first page.