New Website
The final project has taken its first few steps and is shaping up to be quite something.
The old website is still there and accessible through the main screen at the bottom. (It would have been a shame to just delete it)
If you decide to visit the website, please read everything on the main page, it's kinda important.