New Character Page
The Character page has been added with a small description of the person is and the role they play in the show.
A favicon has been added. (May change later).
Added an Image to the story page as well as some more text.
Changed some properties in the style sheet to make the general layout consistent.
Changed "88" to "Patient 88".